refactor: split automation session internals
