+++ crates/blinc_app/src/automation_session/mod.rs
@@ -26,7 +26,9 @@ use crate::windowed::{
 };
 
 mod context;
+mod interactions;
 mod runner;
+mod runtime;
 mod types;
 
 use context::{
@@ -231,247 +233,6 @@ where
         session
     }
 
-    pub fn click(&mut self, locator: AutomationLocator) -> AutomationResult<()> {
-        self.record_command("click", &locator, None);
-        let resolved = self.resolve_target(&locator)?;
-        let target = resolved.target.clone();
-        let Some((local_x, local_y, mouse_x, mouse_y)) = self.target_center(&resolved) else {
-            return Err(self.fail(
-                "target_not_interactable",
-                "click target has no bounds",
-                target,
-                None,
-            ));
-        };
-        let overlays = self.ctx.overlay_manager();
-        let overlay_occludes_target = overlays.has_visible_overlays();
-        let overlay_blocks_background = overlays.has_blocking_overlay();
-        let overlay_handles_backdrop =
-            overlay_blocks_background || overlays.has_dismissable_overlay();
-        if overlay_occludes_target {
-            let overlay_bounds = overlays.get_visible_overlay_bounds();
-            let overlay_layer_id = self.overlay_layer_id();
-            let point_in_overlay_bounds =
-                point_in_overlay_bounds(&overlay_bounds, mouse_x, mouse_y);
-            if (overlay_blocks_background && !point_in_overlay_bounds)
-                || (overlay_handles_backdrop && overlays.is_backdrop_click(mouse_x, mouse_y))
-                || (point_in_overlay_bounds
-                    && !self.target_matches_overlay_hit(
-                        resolved.node_id,
-                        mouse_x,
-                        mouse_y,
-                        &overlay_bounds,
-                        overlay_layer_id,
-                    ))
-            {
-                return Err(self.fail(
-                    "target_blocked_by_overlay",
-                    "click target is occluded by an active overlay",
-                    target,
-                    None,
-                ));
-            }
-        }
-        blinc_layout::widgets::blur_all_text_inputs();
-        let dispatched = self.dispatch_runtime_event(
-            &resolved,
-            ProgrammaticElementEvent::Click {
-                x: local_x,
-                y: local_y,
-            },
-        );
-        if !dispatched {
-            return Err(self.fail(
-                "target_not_interactable",
-                "click did not dispatch any runtime events",
-                target,
-                None,
-            ));
-        }
-        self.after_interaction();
-        Ok(())
-    }
-
-    pub fn click_at(&mut self, mouse_x: f32, mouse_y: f32) -> AutomationResult<()> {
-        self.recording
-            .record_trace_entry(TraceEntryKind::Command(TraceCommandRecord {
-                name: "click".to_string(),
-                target: None,
-                payload: Some(format!("x={mouse_x},y={mouse_y}")),
-            }));
-
-        let overlays = self.ctx.overlay_manager();
-        let overlay_occludes_target = overlays.has_visible_overlays();
-        let overlay_blocks_background = overlays.has_blocking_overlay();
-        let overlay_handles_backdrop =
-            overlay_blocks_background || overlays.has_dismissable_overlay();
-        let overlay_bounds = overlay_occludes_target
-            .then(|| overlays.get_visible_overlay_bounds())
-            .unwrap_or_default();
-        let overlay_layer_id = overlay_occludes_target
-            .then(|| self.overlay_layer_id())
-            .flatten();
-        let point_in_overlay_bounds = point_in_overlay_bounds(&overlay_bounds, mouse_x, mouse_y);
-        if overlay_occludes_target {
-            if overlay_handles_backdrop && overlays.handle_click_at(mouse_x, mouse_y) {
-                self.after_interaction();
-                return Ok(());
-            }
-            if (overlay_blocks_background && !point_in_overlay_bounds)
-                || (overlay_handles_backdrop && overlays.is_backdrop_click(mouse_x, mouse_y))
-            {
-                return Err(self.fail(
-                    "target_blocked_by_overlay",
-                    "click coordinates are occluded by an active overlay",
-                    None,
-                    None,
-                ));
-            }
-        }
-
-        let hit = if overlay_occludes_target {
-            self.ctx.event_router.hit_test_with_occlusion(
-                &self.tree,
-                mouse_x,
-                mouse_y,
-                &overlay_bounds,
-                overlay_layer_id,
-            )
-        } else {
-            self.ctx.event_router.hit_test(&self.tree, mouse_x, mouse_y)
-        };
-        let Some(hit) = hit else {
-            return Err(self.fail(
-                if overlay_occludes_target && (point_in_overlay_bounds || overlay_blocks_background)
-                {
-                    "target_blocked_by_overlay"
-                } else {
-                    "target_not_interactable"
-                },
-                if overlay_occludes_target && (point_in_overlay_bounds || overlay_blocks_background)
-                {
-                    "click coordinates are occluded by an active overlay"
-                } else {
-                    "click coordinates did not hit any element"
-                },
-                None,
-                None,
-            ));
-        };
-
-        blinc_layout::widgets::blur_all_text_inputs();
-        let resolved = ResolvedTarget {
-            node_id: hit.node,
-            target: self.tree.element_registry().get_id(hit.node),
-        };
-        if !self.dispatch_runtime_event(
-            &resolved,
-            ProgrammaticElementEvent::Click {
-                x: hit.local_x,
-                y: hit.local_y,
-            },
-        ) {
-            return Err(self.fail(
-                "target_not_interactable",
-                "click coordinates did not dispatch any runtime events",
-                resolved.target,
-                None,
-            ));
-        }
-        self.after_interaction();
-        Ok(())
-    }
-
-    pub fn fill(&mut self, locator: AutomationLocator, value: &str) -> AutomationResult<()> {
-        self.record_command("fill", &locator, Some(redacted_trace_value(value)));
-        let resolved = self.resolve_target(&locator)?;
-        self.ensure_target_is_unoccluded(&resolved, "fill")?;
-        self.ensure_target_focused(&resolved)?;
-        let select_all_key = parse_key("A").ok_or_else(|| {
-            self.fail(
-                "internal_error",
-                "could not parse select-all key",
-                resolved.target.clone(),
-                None,
-            )
-        })?;
-        let backspace_key = parse_key("Backspace").ok_or_else(|| {
-            self.fail(
-                "internal_error",
-                "could not parse backspace key",
-                resolved.target.clone(),
-                None,
-            )
-        })?;
-        self.dispatch_key_event(&resolved, select_all_key.key_code, select_all_modifiers())?;
-        self.dispatch_key_event(&resolved, backspace_key.key_code, 0)?;
-        for ch in value.chars() {
-            self.dispatch_text_input_event(&resolved, ch, 0)?;
-        }
-        self.after_interaction();
-        Ok(())
-    }
-
-    pub fn assert_exists(&mut self, locator: AutomationLocator) -> AutomationResult<()> {
-        self.record_command("assert_exists", &locator, None);
-        match self.resolve_target(&locator) {
-            Ok(resolved) => {
-                self.record_assertion(
-                    "assert_exists",
-                    true,
-                    resolved.target.as_deref(),
-                    None,
-                    None,
-                );
-                Ok(())
-            }
-            Err(mut failure) => {
-                failure.trace_sequence = Some(self.record_assertion(
-                    "assert_exists",
-                    false,
-                    failure.target.as_deref(),
-                    None,
-                    Some(locator.describe()),
-                ));
-                Err(failure)
-            }
-        }
-    }
-
-    pub fn assert_text_contains(
-        &mut self,
-        locator: AutomationLocator,
-        expected: &str,
-    ) -> AutomationResult<()> {
-        self.record_command("assert_text_contains", &locator, Some(expected.to_string()));
-        let resolved = self.resolve_target(&locator)?;
-        let actual = subtree_text(&self.tree, resolved.node_id).unwrap_or_default();
-        if text_contains(&actual, expected) {
-            self.record_assertion(
-                "assert_text_contains",
-                true,
-                resolved.target.as_deref(),
-                Some(actual),
-                Some(expected.to_string()),
-            );
-            Ok(())
-        } else {
-            let trace_sequence = self.record_assertion(
-                "assert_text_contains",
-                false,
-                resolved.target.as_deref(),
-                Some(actual.clone()),
-                Some(expected.to_string()),
-            );
-            Err(AutomationFailure {
-                code: "assertion_failed".to_string(),
-                message: format!("expected text containing {expected:?}, got {actual:?}"),
-                target: resolved.target,
-                trace_sequence: Some(trace_sequence),
-            })
-        }
-    }
-
     pub fn export_recording(&self) -> RecordingExport {
         self.recording.export()
     }
@@ -493,730 +254,6 @@ where
     ) -> Option<blinc_layout::element::ElementBounds> {
         let node_id = self.tree.query_by_id(id)?;
         self.tree.get_absolute_bounds(node_id)
-    }
-
-    pub fn press(&mut self, key: &str) -> AutomationResult<()> {
-        self.recording
-            .record_trace_entry(TraceEntryKind::Command(TraceCommandRecord {
-                name: "press".to_string(),
-                target: Some(key.to_string()),
-                payload: None,
-            }));
-        let parsed = parse_key(key).ok_or_else(|| {
-            self.fail(
-                "unsupported_key",
-                &format!("unsupported key {key:?}"),
-                None,
-                None,
-            )
-        })?;
-        if parsed.key_code == 27 && self.ctx.overlay_manager().handle_escape() {
-            self.after_interaction();
-            return Ok(());
-        }
-        let Some(node_id) = self.ctx.event_router.focused() else {
-            return Err(self.fail(
-                "focus_required",
-                "press requires a focused element",
-                None,
-                None,
-            ));
-        };
-        let resolved = ResolvedTarget {
-            node_id,
-            target: self.tree.element_registry().get_id(node_id),
-        };
-        self.ensure_target_is_unoccluded(&resolved, "press")?;
-        self.dispatch_key_down_event(&resolved, parsed.key_code, parsed.modifiers)?;
-        if let Some(text) = parsed.text.filter(|_| parsed.modifiers & 0b1010 == 0) {
-            self.dispatch_text_input_event(&resolved, text, parsed.modifiers)?;
-        }
-        self.dispatch_key_up_event(&resolved, parsed.key_code, parsed.modifiers)?;
-        self.after_interaction();
-        Ok(())
-    }
-
-    pub fn scroll(
-        &mut self,
-        locator: Option<AutomationLocator>,
-        dx: f32,
-        dy: f32,
-    ) -> AutomationResult<()> {
-        self.recording
-            .record_trace_entry(TraceEntryKind::Command(TraceCommandRecord {
-                name: "scroll".to_string(),
-                target: locator.as_ref().map(AutomationLocator::describe),
-                payload: Some(format!("dx={dx},dy={dy}")),
-            }));
-        let resolved = match locator {
-            Some(locator) => self.resolve_target(&locator)?,
-            None => {
-                let Some(node_id) = self.ctx.event_router.focused() else {
-                    return Err(self.fail(
-                        "scroll_target_required",
-                        "scroll without an id requires a focused element",
-                        None,
-                        None,
-                    ));
-                };
-                ResolvedTarget {
-                    node_id,
-                    target: None,
-                }
-            }
-        };
-        let Some(bounds) = self.tree.get_absolute_bounds(resolved.node_id) else {
-            return Err(self.fail(
-                "target_not_interactable",
-                "scroll target has no bounds",
-                resolved.target,
-                None,
-            ));
-        };
-        let local_x = bounds.width * 0.5;
-        let local_y = bounds.height * 0.5;
-        let mouse_x = bounds.x + local_x;
-        let mouse_y = bounds.y + local_y;
-        let overlays = self.ctx.overlay_manager();
-        let overlay_occludes_target = overlays.has_visible_overlays();
-        let overlay_blocks_background_scroll = overlays.has_blocking_overlay();
-        if overlay_occludes_target {
-            let overlay_bounds = overlays.get_visible_overlay_bounds();
-            let overlay_layer_id = self.overlay_layer_id();
-            let point_in_overlay_bounds =
-                point_in_overlay_bounds(&overlay_bounds, mouse_x, mouse_y);
-            if (overlay_blocks_background_scroll && !point_in_overlay_bounds)
-                || (point_in_overlay_bounds
-                    && !self.target_matches_overlay_hit(
-                        resolved.node_id,
-                        mouse_x,
-                        mouse_y,
-                        &overlay_bounds,
-                        overlay_layer_id,
-                    ))
-            {
-                return Err(self.fail(
-                    "target_blocked_by_overlay",
-                    "scroll target is occluded by an active overlay",
-                    resolved.target,
-                    None,
-                ));
-            }
-        }
-        let overlay_effect = self.handle_overlay_scroll(dy);
-        let dispatched =
-            self.dispatch_runtime_event(&resolved, ProgrammaticElementEvent::Scroll { dx, dy });
-        if !dispatched && !overlay_effect {
-            return Err(self.fail(
-                "target_not_interactable",
-                &format!(
-                    "scroll target at ({:.1}, {:.1}, {:.1}, {:.1}) did not dispatch",
-                    bounds.x, bounds.y, bounds.width, bounds.height
-                ),
-                resolved.target,
-                None,
-            ));
-        }
-        self.after_scroll_interaction();
-        Ok(())
-    }
-
-    pub fn tick_frames(&mut self, frames: u32) -> Result<()> {
-        if frames == 0 {
-            return Ok(());
-        }
-        let probe_every = self.runtime_cfg.probe_every_frames.max(1);
-        let mut current_time = self.last_frame_time_ms;
-        for frame_index in 0..frames {
-            if let Ok(animations) = self.ctx.animations.lock() {
-                let _ = animations.tick();
-            }
-            current_time = current_time.saturating_add(self.frame_step_ms());
-            let should_sample = (frame_index + 1) % probe_every == 0 || frame_index + 1 == frames;
-            self.advance_runtime_frame(current_time, should_sample, false);
-        }
-        Ok(())
-    }
-
-    pub fn write_snapshot_to_path(&self, path: &std::path::Path) -> Result<()> {
-        let Some(snapshot) = self.latest_snapshot.as_ref() else {
-            bail!("no snapshot captured yet");
-        };
-        if let Some(parent) = path.parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(parent)?;
-            }
-        }
-        std::fs::write(path, serde_json::to_string_pretty(snapshot)?)?;
-        self.recording
-            .record_trace_entry(TraceEntryKind::Artifact(TraceArtifactRecord {
-                kind: "snapshot_export".to_string(),
-                path: Some(path.display().to_string()),
-                message: Some("wrote snapshot".to_string()),
-            }));
-        Ok(())
-    }
-
-    pub fn write_trace_to_path(&self, path: &std::path::Path) -> Result<()> {
-        if let Some(parent) = path.parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(parent)?;
-            }
-        }
-        let artifact = TraceArtifactRecord {
-            kind: "trace_export".to_string(),
-            path: Some(path.display().to_string()),
-            message: Some("wrote trace".to_string()),
-        };
-        let artifact_entry = self
-            .recording
-            .prepare_trace_entry(TraceEntryKind::Artifact(artifact.clone()));
-        let mut export = self.export_recording();
-        if let Some(entry) = artifact_entry.clone() {
-            export.trace_entries.push(entry);
-        }
-        std::fs::write(path, serde_json::to_string_pretty(&export)?)?;
-        if let Some(entry) = artifact_entry {
-            let _ = self.recording.append_trace_entry(entry);
-        }
-        Ok(())
-    }
-
-    fn resolve_target(&mut self, locator: &AutomationLocator) -> AutomationResult<ResolvedTarget> {
-        match locator {
-            AutomationLocator::Id(id) => {
-                let trace_sequence =
-                    self.recording
-                        .record_trace_entry(TraceEntryKind::LocatorResolution(
-                            TraceLocatorResolution {
-                                query: format!("id={id:?}"),
-                                matched_target: self.tree.query_by_id(id).map(|_| id.clone()),
-                                candidate_targets: self
-                                    .tree
-                                    .query_by_id(id)
-                                    .map(|_| vec![id.clone()])
-                                    .unwrap_or_default(),
-                                failure_reason: self
-                                    .tree
-                                    .query_by_id(id)
-                                    .is_none()
-                                    .then(|| "no_match".to_string()),
-                            },
-                        ));
-                let Some(node_id) = self.tree.query_by_id(id) else {
-                    return Err(AutomationFailure {
-                        code: "locator_not_found".to_string(),
-                        message: format!("no element found for id {id:?}"),
-                        target: Some(id.clone()),
-                        trace_sequence: Some(trace_sequence),
-                    });
-                };
-                Ok(ResolvedTarget {
-                    node_id,
-                    target: Some(id.clone()),
-                })
-            }
-            AutomationLocator::Semantic(locator) => {
-                let resolution = resolve_semantic_locator(&self.tree, locator);
-                let Some(node_id) = resolution.matched_node_id else {
-                    return Err(AutomationFailure {
-                        code: resolution
-                            .failure_reason
-                            .clone()
-                            .unwrap_or_else(|| "locator_resolution_failed".to_string()),
-                        message: format!("semantic locator failed: {}", resolution.query),
-                        target: resolution.matched_target.clone(),
-                        trace_sequence: None,
-                    });
-                };
-                Ok(ResolvedTarget {
-                    node_id,
-                    target: resolution.matched_target,
-                })
-            }
-        }
-    }
-
-    fn target_center(&self, resolved: &ResolvedTarget) -> Option<(f32, f32, f32, f32)> {
-        let bounds = self.tree.get_absolute_bounds(resolved.node_id)?;
-        let local_x = bounds.width * 0.5;
-        let local_y = bounds.height * 0.5;
-        Some((local_x, local_y, bounds.x + local_x, bounds.y + local_y))
-    }
-
-    fn ensure_target_is_unoccluded(
-        &self,
-        resolved: &ResolvedTarget,
-        action: &str,
-    ) -> AutomationResult<()> {
-        let Some((_, _, mouse_x, mouse_y)) = self.target_center(resolved) else {
-            return Err(self.fail(
-                "target_not_interactable",
-                &format!("{action} target has no bounds"),
-                resolved.target.clone(),
-                None,
-            ));
-        };
-
-        let overlays = self.ctx.overlay_manager();
-        if !overlays.has_visible_overlays() {
-            return Ok(());
-        }
-
-        let overlay_bounds = overlays.get_visible_overlay_bounds();
-        let overlay_layer_id = self.overlay_layer_id();
-        let point_in_overlay_bounds = point_in_overlay_bounds(&overlay_bounds, mouse_x, mouse_y);
-        if (overlays.has_blocking_overlay() && !point_in_overlay_bounds)
-            || (point_in_overlay_bounds
-                && !self.target_matches_overlay_hit(
-                    resolved.node_id,
-                    mouse_x,
-                    mouse_y,
-                    &overlay_bounds,
-                    overlay_layer_id,
-                ))
-        {
-            return Err(self.fail(
-                "target_blocked_by_overlay",
-                &format!("{action} target is occluded by an active overlay"),
-                resolved.target.clone(),
-                None,
-            ));
-        }
-
-        Ok(())
-    }
-
-    fn overlay_layer_id(&self) -> Option<blinc_layout::tree::LayoutNodeId> {
-        self.tree
-            .query_by_id(blinc_layout::widgets::overlay::OVERLAY_LAYER_ID)
-    }
-
-    fn target_matches_overlay_hit(
-        &self,
-        node_id: blinc_layout::tree::LayoutNodeId,
-        x: f32,
-        y: f32,
-        overlay_bounds: &[(f32, f32, f32, f32)],
-        overlay_layer_id: Option<blinc_layout::tree::LayoutNodeId>,
-    ) -> bool {
-        let Some(hit) = self.ctx.event_router.hit_test_with_occlusion(
-            &self.tree,
-            x,
-            y,
-            overlay_bounds,
-            overlay_layer_id,
-        ) else {
-            return false;
-        };
-
-        if hit.node == node_id || hit.ancestors.contains(&node_id) {
-            return true;
-        }
-
-        let root_id = self.tree.root();
-        self.tree
-            .element_registry()
-            .ancestors(node_id)
-            .into_iter()
-            .any(|ancestor| Some(ancestor) != root_id && ancestor == hit.node)
-    }
-
-    fn handle_overlay_scroll(&mut self, delta_y: f32) -> bool {
-        let updated = self.ctx.overlay_manager().handle_scroll(delta_y);
-        if updated {
-            self.sync_overlay_scroll_offsets();
-        }
-        updated
-    }
-
-    fn sync_overlay_scroll_offsets(&mut self) {
-        let overlays = self.ctx.overlay_manager();
-        for (element_id, offset_y) in overlays.get_scroll_offsets() {
-            if let Some(node_id) = self.tree.query_by_id(&element_id) {
-                self.tree.set_scroll_offset(node_id, 0.0, offset_y);
-            }
-        }
-    }
-
-    fn ensure_target_focused(&mut self, resolved: &ResolvedTarget) -> AutomationResult<()> {
-        if self.ctx.event_router.focused() == Some(resolved.node_id) {
-            return Ok(());
-        }
-
-        if sync_focus_node_to_runtime(
-            &mut self.tree,
-            &mut self.ctx.event_router,
-            Some(resolved.node_id),
-        ) {
-            return Ok(());
-        }
-
-        Err(self.fail(
-            "focus_dispatch_failed",
-            "target could not be focused through the runtime",
-            resolved.target.clone(),
-            None,
-        ))
-    }
-
-    fn dispatch_key_event(
-        &mut self,
-        resolved: &ResolvedTarget,
-        key: u32,
-        modifiers: u8,
-    ) -> AutomationResult<()> {
-        self.dispatch_key_down_event(resolved, key, modifiers)?;
-        self.dispatch_key_up_event(resolved, key, modifiers)
-    }
-
-    fn dispatch_key_down_event(
-        &mut self,
-        resolved: &ResolvedTarget,
-        key: u32,
-        modifiers: u8,
-    ) -> AutomationResult<()> {
-        if self.dispatch_runtime_event(
-            resolved,
-            ProgrammaticElementEvent::KeyDown { key, modifiers },
-        ) {
-            Ok(())
-        } else {
-            Err(self.fail(
-                "key_dispatch_failed",
-                "key press did not dispatch to the runtime",
-                resolved.target.clone(),
-                None,
-            ))
-        }
-    }
-
-    fn dispatch_key_up_event(
-        &mut self,
-        resolved: &ResolvedTarget,
-        key: u32,
-        modifiers: u8,
-    ) -> AutomationResult<()> {
-        if self.dispatch_runtime_event(resolved, ProgrammaticElementEvent::KeyUp { key, modifiers })
-        {
-            Ok(())
-        } else {
-            Err(self.fail(
-                "key_dispatch_failed",
-                "key release did not dispatch to the runtime",
-                resolved.target.clone(),
-                None,
-            ))
-        }
-    }
-
-    fn dispatch_text_input_event(
-        &mut self,
-        resolved: &ResolvedTarget,
-        text: char,
-        modifiers: u8,
-    ) -> AutomationResult<()> {
-        if self.dispatch_runtime_event(
-            resolved,
-            ProgrammaticElementEvent::TextInput { text, modifiers },
-        ) {
-            Ok(())
-        } else {
-            Err(self.fail(
-                "text_input_dispatch_failed",
-                "text input did not dispatch to the runtime",
-                resolved.target.clone(),
-                None,
-            ))
-        }
-    }
-
-    fn dispatch_runtime_event(
-        &mut self,
-        resolved: &ResolvedTarget,
-        event: ProgrammaticElementEvent,
-    ) -> bool {
-        dispatch_programmatic_event_to_node(
-            &mut self.tree,
-            &mut self.ctx.event_router,
-            resolved.node_id,
-            event,
-        )
-    }
-
-    fn apply_pending_runtime_requests(&mut self) -> bool {
-        drain_programmatic_runtime_requests(
-            &mut self.tree,
-            &mut self.ctx.event_router,
-            &self.pending_focus_changes,
-            &self.pending_scroll_requests,
-            &self.pending_programmatic_events,
-        )
-    }
-
-    fn after_interaction(&mut self) {
-        self.mark_visible_overlay_content_dirty();
-        let current_time = self.last_frame_time_ms.saturating_add(self.frame_step_ms());
-        self.advance_runtime_frame(current_time, true, true);
-    }
-
-    fn after_scroll_interaction(&mut self) {
-        let current_time = self.last_frame_time_ms.saturating_add(self.frame_step_ms());
-        self.advance_runtime_frame(current_time, true, false);
-    }
-
-    fn advance_runtime_frame(
-        &mut self,
-        current_time: u64,
-        capture_snapshot: bool,
-        force_rebuild: bool,
-    ) {
-        self.prepare_runtime_frame(current_time);
-
-        let _ = self.apply_pending_runtime_requests();
-        let _ = self.tree.tick_scroll_physics(current_time);
-        self.tree.process_pending_scroll_refs();
-        self.apply_stateful_updates();
-        self.rebuild_runtime_tree(force_rebuild);
-        self.finalize_runtime_frame(current_time);
-
-        if let Some(focused_id) = BlincContextState::try_get().and_then(|ctx| ctx.focused_element())
-        {
-            let runtime_focused_id = self
-                .ctx
-                .event_router
-                .focused()
-                .and_then(|node_id| self.tree.element_registry().get_id(node_id));
-            if runtime_focused_id.as_deref() != Some(focused_id.as_str()) {
-                sync_focus_to_runtime(
-                    &mut self.tree,
-                    &mut self.ctx.event_router,
-                    Some(&focused_id),
-                );
-            }
-        } else {
-            let text_focus = blinc_layout::widgets::text_input::focused_text_input_node_id()
-                .or_else(blinc_layout::widgets::text_input::focused_text_area_node_id);
-            if text_focus != self.ctx.event_router.focused() {
-                self.ctx.event_router.set_focus(text_focus);
-            }
-        }
-        sync_context_focus_from_runtime(&self.tree, &self.ctx.event_router);
-        if capture_snapshot {
-            self.capture_snapshot();
-        }
-    }
-
-    fn prepare_runtime_frame(&mut self, current_time: u64) {
-        if blinc_layout::widgets::take_needs_css_reparse() {
-            self.ctx.reparse_css();
-        }
-
-        self.render_state.process_global_motion_exit_starts();
-        self.render_state.process_global_motion_exit_cancels();
-        self.render_state.process_global_motion_starts();
-        self.render_state.sync_shared_motion_states();
-
-        self.ctx.prepare_windowless_frame(current_time);
-        let overlay_content_dirty = self.ctx.overlay_manager().is_dirty();
-        if overlay_content_dirty {
-            if let Some(overlay_node_id) = self
-                .element_registry
-                .get(blinc_layout::widgets::overlay::OVERLAY_LAYER_ID)
-            {
-                let overlay_content = self.ctx.overlay_manager().build_overlay_layer();
-                blinc_layout::queue_subtree_rebuild(overlay_node_id, overlay_content);
-            }
-            let _ = self.ctx.overlay_manager().take_dirty();
-        }
-    }
-
-    fn apply_stateful_updates(&mut self) {
-        let has_stateful_updates = blinc_layout::take_needs_redraw();
-        let has_pending_rebuilds = blinc_layout::has_pending_subtree_rebuilds();
-        if !has_stateful_updates && !has_pending_rebuilds {
-            return;
-        }
-
-        let prop_updates = blinc_layout::take_pending_prop_updates();
-        for (node_id, props) in &prop_updates {
-            self.tree
-                .update_render_props(*node_id, |render_props| *render_props = props.clone());
-        }
-
-        if self.tree.process_pending_subtree_rebuilds() {
-            self.tree.apply_stylesheet_layout_overrides();
-            self.tree.compute_layout(
-                self.runtime_cfg.width as f32,
-                self.runtime_cfg.height as f32,
-            );
-            self.render_state.begin_stable_motion_frame();
-            self.tree
-                .initialize_motion_animations(&mut self.render_state);
-            self.render_state.end_stable_motion_frame();
-            self.render_state.process_global_motion_replays();
-            self.tree.start_all_css_animations();
-        }
-    }
-
-    fn rebuild_runtime_tree(&mut self, force_rebuild: bool) {
-        let needs_rebuild = force_rebuild
-            || self.tree.needs_rebuild()
-            || self.ref_dirty_flag.swap(false, Ordering::SeqCst)
-            || blinc_layout::widgets::take_needs_rebuild();
-        let needs_relayout = force_rebuild || blinc_layout::widgets::take_needs_relayout();
-
-        self.render_state.begin_stable_motion_frame();
-        if !needs_rebuild {
-            self.tree
-                .initialize_motion_animations(&mut self.render_state);
-            self.render_state.end_stable_motion_frame();
-            return;
-        }
-
-        blinc_layout::reset_call_counters();
-        blinc_layout::clear_stateful_base_updaters();
-        blinc_layout::click_outside::clear_click_outside_handlers();
-        self.render_state.reset_stable_motions_for_rebuild();
-
-        let user_ui = (self.ui_builder)(&mut self.ctx);
-        let ui = self.ctx.compose_runtime_ui(user_ui);
-
-        if let Some(ref stylesheet) = self.ctx.stylesheet {
-            self.tree.set_stylesheet_arc(stylesheet.clone());
-        }
-
-        if needs_relayout {
-            let mut tree = blinc_layout::RenderTree::from_element_with_registry(
-                &ui,
-                Arc::clone(&self.element_registry),
-            );
-            tree.set_animations(&self.ctx.animations);
-            tree.set_css_anim_store(Arc::clone(&self.css_anim_store));
-            tree.set_scale_factor(self.ctx.scale_factor as f32);
-            if let Some(ref stylesheet) = self.ctx.stylesheet {
-                tree.set_stylesheet_arc(stylesheet.clone());
-            }
-            tree.apply_all_stylesheet_styles();
-            tree.compute_layout(
-                self.runtime_cfg.width as f32,
-                self.runtime_cfg.height as f32,
-            );
-            tree.transfer_scroll_state_from(&self.tree);
-            tree.initialize_motion_animations(&mut self.render_state);
-            self.render_state.end_stable_motion_frame();
-            self.render_state.process_global_motion_replays();
-            tree.start_all_css_animations();
-            self.tree = tree;
-        } else {
-            match self.tree.incremental_update(&ui) {
-                UpdateResult::NoChanges | UpdateResult::VisualOnly => {
-                    self.tree
-                        .initialize_motion_animations(&mut self.render_state);
-                    self.render_state.end_stable_motion_frame();
-                }
-                UpdateResult::LayoutChanged => {
-                    self.tree.apply_stylesheet_layout_overrides();
-                    self.tree.compute_layout(
-                        self.runtime_cfg.width as f32,
-                        self.runtime_cfg.height as f32,
-                    );
-                    self.tree.restore_bound_scroll_ref_offsets();
-                    self.tree
-                        .initialize_motion_animations(&mut self.render_state);
-                    self.render_state.end_stable_motion_frame();
-                }
-                UpdateResult::ChildrenChanged => {
-                    self.tree.apply_stylesheet_base_styles();
-                    self.tree.apply_stylesheet_layout_overrides();
-                    self.tree.compute_layout(
-                        self.runtime_cfg.width as f32,
-                        self.runtime_cfg.height as f32,
-                    );
-                    self.tree.restore_bound_scroll_ref_offsets();
-                    self.tree
-                        .initialize_motion_animations(&mut self.render_state);
-                    self.render_state.end_stable_motion_frame();
-                    self.render_state.process_global_motion_replays();
-                    self.tree.start_all_css_animations();
-                }
-            }
-        }
-
-        self.ctx.finish_runtime_rebuild();
-        self.sync_overlay_scroll_offsets();
-        self.tree.process_pending_scroll_refs();
-    }
-
-    fn mark_visible_overlay_content_dirty(&self) {
-        self.ctx.overlay_manager().mark_content_dirty();
-    }
-
-    fn finalize_runtime_frame(&mut self, current_time: u64) {
-        self.render_state.process_global_motion_exit_cancels();
-        self.render_state.process_global_motion_exit_starts();
-        self.render_state.process_global_motion_starts();
-        let _ = self.render_state.tick(current_time);
-
-        let dt_ms = frame_delta_ms(current_time, self.last_frame_time_ms);
-        let css_active = {
-            let store = self.tree.css_anim_store();
-            let mut animations = store.lock().unwrap();
-            let (animating, transitioning) = animations.tick(dt_ms);
-            drop(animations);
-            animating || transitioning || self.tree.css_has_active()
-        };
-        self.last_frame_time_ms = current_time;
-        self.render_state.sync_shared_motion_states();
-        let _ = blinc_theme::ThemeState::get().tick();
-
-        if self.tree.stylesheet().is_some() {
-            let state_changed = self
-                .tree
-                .apply_stylesheet_state_styles(&self.ctx.event_router);
-            if state_changed {
-                self.tree.compute_layout(
-                    self.runtime_cfg.width as f32,
-                    self.runtime_cfg.height as f32,
-                );
-            }
-        }
-        sync_context_focus_from_runtime(&self.tree, &self.ctx.event_router);
-
-        if css_active || !self.tree.css_transitions_empty() {
-            self.tree.apply_all_css_animation_props();
-            self.tree.apply_all_css_transition_props();
-            if self.tree.apply_animated_layout_props() {
-                self.tree.compute_layout(
-                    self.runtime_cfg.width as f32,
-                    self.runtime_cfg.height as f32,
-                );
-            }
-        }
-        self.sync_overlay_scroll_offsets();
-        self.tree.process_pending_scroll_refs();
-    }
-
-    fn capture_snapshot(&mut self) {
-        let hovered_nodes = self
-            .ctx
-            .event_router
-            .hovered_nodes()
-            .collect::<std::collections::HashSet<_>>();
-        let snapshot = capture_tree_snapshot(
-            &self.tree,
-            self.ctx.event_router.focused(),
-            &hovered_nodes,
-            self.runtime_cfg.width,
-            self.runtime_cfg.height,
-        );
-        let snapshot = to_tree_snapshot(snapshot);
-        self.recording.record_snapshot(snapshot.clone());
-        self.latest_snapshot = Some(snapshot);
-    }
-
-    fn frame_step_ms(&self) -> u64 {
-        u64::from(self.runtime_cfg.tick_ms.max(1))
     }
 
     fn record_command(
